Setup code highlighting
Enable code highlighting for Python and bash code. Provide suitable Vue
components, and configure Nuxt Content to highlight code blocks in
markdown.

Deviate from the style guide in that we are using a light colored
background for code samples, instead of the specified warm-black one, to
work around the inability to specify different syntax highlighting
themes for inline and block code.