Prose: Make headlines as long as body text
When using `max-w-prose`, headlines were longer than the body text.
max-w-prose limits the amount of characters in a line, which is results
in wider text for headlines in a bigger font than for body text. Specify
the width in pixels for a font-size-agnostic description.

Now that we're breaking headlines more often, add `text-wrap: balance`
to reduce orphans. (Such as "Latch-\nUp").