Remove flowbite from the main navigation bar
The main navigation bar (on the top) uses the Collapsible components
from Flowbite to provide actions for the hamburger menu. It turns out
that Flowbite is a bit annoying to get right with SSR, so let's just
remove it and rely in pure Vue concepts for something as trivial as a
collapsible menu bar.

While at it, also
* Follow the style guide more closely and display a close button
  whenever the menu is open.
* Reduce the amount of executed JavaScript.
* Use icons directly from the icon library.